Refactor GPU code
